Changes I made to fix input problems in i/o structure of this program and a cascade of related changes when I worked through the code.
1.  I cleaned up some inconsistencies in the database interface.  Basically
I converted to a pure OOP version using the Datascope_Handle object.

2. Some minor rearranging of the main try/catch blocks in pwstack.C

3.  The biggest change is I'm changing the basic flow of pwstack.C in a fundamental
way.  It is now driven by a loop over a 2d "GCLgrid" object produced by a constructor.
A GCLgrid is the subject of a paper Fan and I are trying to get resubmitted
after an initial review.  We can discuss more later, but basically it is a way
to define a georeferenced grid of points, which is what is needed here.
Look at the top of the main loop to see the primary changes.

4.  pwstack_ensemble argument list has changed because I decided to push some
parameters through the ensemble's metadata area.  Kind of a slow way to pass
arguments to a function, but I think it may be clearer and allows additions
later without messing with the arg list.  Function is not called millions
of times so should not be a huge issue.

5.  Did the output process correctly in te pwstack_ensemble function.  It
now at least doesn't contain the fundamental error it had before.

6.  Changes in the schema definition.